arquitetura de movies
