+++ src/movies/controllers/movies.controller.spec.ts
@@ -1,8 +1,8 @@
-import { getModelToken } from '@nestjs/mongoose';
 import { Test, TestingModule } from '@nestjs/testing';
-import { Model } from 'mongoose';
-import { Movie } from './movie';
-import { MovieService } from './movie.service';
+import { Pages } from '../model/pages';
+import { Movie } from '../model/movie';
+import { MovieService } from '../services/movie.service';
+import { MoviesController } from './movies.controller';
 
 const movie: Movie = {
   _id: '1',
@@ -101,6 +101,43 @@ const movies: Movie[] = [
     year: 2020,
     imdb: { rating: 6, votes: 200, id: 1111 },
     type: 'Movie',
+    tomatoes: {
+      viewer: { rating: 6, numReviews: 150, meter: 83 },
+      dvd: new Date('2021-10-11T00:00:00.000+00:00'),
+      lastUpdated: new Date('2022-07-11T00:00:00.000+00:00'),
+      rotten: 1,
+      fresh: 12,
+      critic: { rating: 7.2, numReviews: 13, meter: 92 },
+    },
+    dvd: new Date('2021-10-11T00:00:00.000+00:00'),
+    fresh: 12,
+    production: 'Warner Brothers',
+    rotten: 1,
+    lastUpdated: new Date('2022-8-11T00:00:00.000+00:00'),
+    website: 'www.filme.com',
+    writers: ['Tarantino'],
+  },
+  {
+    _id: '3',
+    plot: 'era uma vez uma vez...',
+    genres: ['Ação', 'Aventura'],
+    runtime: 120,
+    cast: ['Brad Pitt', 'Tom Cruise', 'Jennifer Lawrence'],
+    num_mflix_comments: 5,
+    poster: 'poster.png',
+    title: 'Tome Bala',
+    fullplot: 'Max que matou o Nilo.',
+    metacritic: 5,
+    languages: ['Portuguese', 'English'],
+    countries: ['Brazil'],
+    realeased: new Date('2020-07-11T00:00:00.000+00:00'),
+    directors: ['Tarantino', 'Christhofer Nolan'],
+    rated: 'Good',
+    awards: { wins: 1, nominations: 4, text: '1 win and 4 nominations' },
+    lastupdadted: '2021-07-11T00:00:00.000+00:00',
+    year: 2020,
+    imdb: { rating: 6, votes: 200, id: 1111 },
+    type: 'Series',
     tomatoes: {
       viewer: { rating: 6, numReviews: 150, meter: 83 },
       dvd: new Date('2021-10-11T00:00:00.000+00:00'),
@@ -237,213 +274,110 @@ const series: Movie[] = [
 
 const count: number = 1;
 
-describe('MovieService', () => {
+describe('MoviesController', () => {
+  let moviesController: MoviesController;
   let movieService: MovieService;
-  let movieModel: Model<Movie>;
-
-  let pagination = {
-    limit: 10,
-    page: 2,
-  };
-  let skip = pagination.limit * (pagination.page - 1);
-
-  const mockMovie = {
-    getMovies: jest.fn().mockResolvedValue(movies),
-    findOne: jest.fn().mockResolvedValue(movies),
-    create: jest.fn().mockResolvedValue(newMovie),
-    update: jest.fn().mockResolvedValue(updatedMovie),
-    delete: jest.fn().mockResolvedValue(undefined),
-    countDocuments: jest.fn().mockResolvedValue(movies.length),
-    skip: jest.fn().mockResolvedValue(skip),
-  };
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
+      controllers: [MoviesController],
       providers: [
         {
           provide: MovieService,
-          useValue: mockMovie,
-        },
-        {
-          provide: getModelToken('Movie'),
-          useValue: mockMovie,
+          useValue: {
+            getAll: jest.fn().mockResolvedValue(movies),
+            findOne: jest.fn().mockResolvedValue(movies),
+            findSeries: jest.fn().mockResolvedValue(series),
+            getMovies: jest.fn().mockResolvedValue(movies),
+            findByMovieId: jest.fn().mockResolvedValue(movie),
+            create: jest.fn().mockResolvedValue(newMovie),
+            update: jest.fn().mockResolvedValue(updatedMovie),
+            delete: jest.fn().mockResolvedValue(undefined),
+            // getCategory: jest.fn().mockResolvedValue(movie),
+            // getDirectors: jest.fn().mockResolvedValue(movie),
+            // getCast: jest.fn().mockResolvedValue(movie),
+            // getLetter: jest.fn().mockResolvedValue(movie),
+            // getByYear: jest.fn().mockResolvedValue(movie),
+          },
         },
       ],
     }).compile();
 
+    moviesController = module.get<MoviesController>(MoviesController);
     movieService = module.get<MovieService>(MovieService);
-    movieModel = module.get<Model<Movie>>(getModelToken('Movie'));
   });
 
   it('should be defined', () => {
+    expect(moviesController).toBeDefined();
     expect(movieService).toBeDefined();
-    expect(movieModel).toBeDefined();
   });
 
   describe('getAll', () => {
     it('Deve retornar catalogo completo da tallosflix', async () => {
       // Act
-      const result = await movieService.getMovies(movie, pagination);
-      const countDocuments = await movieModel.countDocuments(movies);
-      const valueCount = movies.length;
-
+      const result = await moviesController.getAll(movie);
       // Assert
-      expect(skip).toEqual(pagination.limit);
-      expect(countDocuments).toBe(valueCount);
-      expect(result).toEqual(movies);
+      expect(result).toBe(movies);
       expect(typeof result).toEqual('object');
       expect(movieService.getMovies).toHaveBeenCalledTimes(1);
     });
 
-    it('should throw an exception', () => {
-      // Arrange
-      jest
-        .spyOn(movieService, 'getMovies')
-        .mockRejectedValueOnce(new Error('Bad Request'));
-
-      // Assert
-      expect(movieService.getMovies).rejects.toThrowError('Bad Request');
-    });
-  });
-
-  // describe('getAllSeries', () => {
-  //   it('Deve retornar lista de series', async () => {
-  //     // Act
-  //     const result = await movieService.getAllSeries();
-
-  //     // Assert
-  //     expect(result).toEqual(series);
-  //     expect(typeof result).toEqual('object');
-  //     expect(movieService.getAllSeries).toHaveBeenCalledTimes(1);
-  //   });
-
-  //   it('should throw an exception', () => {
-  //     // Arrange
-  //     jest
-  //       .spyOn(movieService, 'getAllSeries')
-  //       .mockRejectedValueOnce(new Error());
-
-  //     // Assert
-  //     expect(movieService.getAllSeries()).rejects.toThrowError();
-  //   });
-  // });
-
-  describe('create', () => {
-    it('Deve criar um novo movie com sucesso', async () => {
-      // Arrange
-      const body: Movie = {
-        _id: '1',
-        plot: 'era uma vez uma vez...',
-        genres: ['Ação', 'Aventura'],
-        runtime: 120,
-        cast: ['Brad Pitt', 'Tom Cruise', 'Jennifer Lawrence'],
-        num_mflix_comments: 5,
-        poster: 'poster.png',
-        title: 'Tome Bala',
-        fullplot: 'Max que matou o Nilo.',
-        metacritic: 5,
-        languages: ['Portuguese', 'English'],
-        countries: ['Brazil'],
-        realeased: new Date('2020-07-11T00:00:00.000+00:00'),
-        directors: ['Tarantino', 'Christhofer Nolan'],
-        rated: 'Good',
-        awards: { wins: 1, nominations: 4, text: '1 win and 4 nominations' },
-        lastupdadted: '2021-07-11T00:00:00.000+00:00',
-        year: 2020,
-        imdb: { rating: 6, votes: 200, id: 1111 },
-        type: 'Movie',
-        tomatoes: {
-          viewer: { rating: 6, numReviews: 150, meter: 83 },
-          dvd: new Date('2021-10-11T00:00:00.000+00:00'),
-          lastUpdated: new Date('2022-07-11T00:00:00.000+00:00'),
-          rotten: 1,
-          fresh: 12,
-          critic: { rating: 7.2, numReviews: 13, meter: 92 },
-        },
-        dvd: new Date('2021-10-11T00:00:00.000+00:00'),
-        fresh: 12,
-        production: 'Warner Brothers',
-        rotten: 1,
-        lastUpdated: new Date('2022-8-11T00:00:00.000+00:00'),
-        website: 'www.filme.com',
-        writers: ['Tarantino'],
-      };
-
+    it('Deve retornar catalogo de filmes de acordo com o type de busca ano', async () => {
       // Act
-      const result = await movieService.create(body);
-
+      const result = await moviesController.findOne({ year: 2020 });
       // Assert
-      expect(result).toEqual(newMovie);
-      expect(movieService.create).toHaveBeenCalledTimes(1);
-      expect(movieService.create).toHaveBeenCalledWith(body);
+      expect(result).toBe(movie);
+      // expect(typeof result).toEqual('object');
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
     });
 
-    it('should throw an exception', () => {
-      // Arrange
-      const body: Movie = {
-        _id: '1',
-        plot: 'era uma vez uma vez...',
-        genres: ['Ação', 'Aventura'],
-        runtime: 120,
-        cast: ['Brad Pitt', 'Tom Cruise', 'Jennifer Lawrence'],
-        num_mflix_comments: 5,
-        poster: 'poster.png',
-        title: 'Tome Bala',
-        fullplot: 'Max que matou o Nilo.',
-        metacritic: 5,
-        languages: ['Portuguese', 'English'],
-        countries: ['Brazil'],
-        realeased: new Date('2020-07-11T00:00:00.000+00:00'),
-        directors: ['Tarantino', 'Christhofer Nolan'],
-        rated: 'Good',
-        awards: { wins: 1, nominations: 4, text: '1 win and 4 nominations' },
-        lastupdadted: '2021-07-11T00:00:00.000+00:00',
-        year: 2020,
-        imdb: { rating: 6, votes: 200, id: 1111 },
-        type: 'Movie',
-        tomatoes: {
-          viewer: { rating: 6, numReviews: 150, meter: 83 },
-          dvd: new Date('2021-10-11T00:00:00.000+00:00'),
-          lastUpdated: new Date('2022-07-11T00:00:00.000+00:00'),
-          rotten: 1,
-          fresh: 12,
-          critic: { rating: 7.2, numReviews: 13, meter: 92 },
-        },
-        dvd: new Date('2021-10-11T00:00:00.000+00:00'),
-        fresh: 12,
-        production: 'Warner Brothers',
-        rotten: 1,
-        lastUpdated: new Date('2022-8-11T00:00:00.000+00:00'),
-        website: 'www.filme.com',
-        writers: ['Tarantino'],
-      };
-
-      jest.spyOn(movieService, 'create').mockRejectedValueOnce(new Error());
-
+    it('Deve retornar catalogo de filmes de acordo com o type de busca diretores', async () => {
+      // Act
+      const result = await moviesController.findOne({
+        directors: ['Tarantino'],
+      });
       // Assert
-      expect(movieService.create(body)).rejects.toThrowError();
+      expect(result).toBe(movie);
+      // expect(typeof result).toEqual('object');
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
+    });
+
+    it('Deve retornar catalogo de filmes de acordo com o type de busca genero', async () => {
+      // Act
+      const result = await moviesController.findOne({ genres: ['Aventura'] });
+      // Assert
+      expect(result).toBe(movie);
+      // expect(typeof result).toEqual('object');
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
+    });
+
+    it('Deve retornar catalogo de filmes de acordo com o type de busca escritores', async () => {
+      // Act
+      const result = await moviesController.findOne({ writers: ['Tarantino'] });
+      // Assert
+      expect(result).toBe(movie);
+      // expect(typeof result).toEqual('object');
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
+    });
+
+    it('Deve retornar catalogo de filmes de acordo com o type de busca elenco', async () => {
+      // Act
+      const result = await moviesController.findOne({ cast: ['Tom Cruise'] });
+      // Assert
+      expect(result).toBe(movie);
+      // expect(typeof result).toEqual('object');
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
+    });
+
+    it('Deve retornar catalogo de filmes de acordo com o type de busca movie', async () => {
+      // Act
+      const result = await moviesController.findOne({ type: 'movie' });
+      // Assert
+      expect(result).toEqual(movie);
+
+      expect(movieService.findByMovieId).toHaveBeenCalledTimes(1);
     });
   });
-
-  // describe('getById', () => {
-  //   it('Deve retornar um movie com sucesso pelo ID', async () => {
-  //     // Act
-  //     const result = await movieService.getById('1');
-
-  //     // Assert
-  //     expect(result).toEqual(movie[0]);
-  //     expect(movieService.getById).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getById).toHaveBeenCalledWith('1');
-  //   });
-
-  //   it('should throw an exception', () => {
-  //     // Arrange
-  //     jest.spyOn(movieService, 'getById').mockRejectedValueOnce(new Error());
-
-  //     // Assert
-  //     expect(movieService.getById('1')).rejects.toThrowError();
-  //   });
-  // });
 
   describe('update', () => {
     it('Deve alterar dados de um movie pelo ID', async () => {
@@ -487,7 +421,7 @@ describe('MovieService', () => {
       };
 
       // Act
-      const result = await movieService.update('1', body);
+      const result = await moviesController.update('1', body);
 
       // Assert
       expect(result).toEqual(updatedMovie);
@@ -538,7 +472,7 @@ describe('MovieService', () => {
       jest.spyOn(movieService, 'update').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(movieService.update(body._id, body)).rejects.toThrowError();
+      expect(moviesController.update(body._id, body)).rejects.toThrowError();
     });
   });
 
@@ -550,7 +484,7 @@ describe('MovieService', () => {
       };
 
       //Act
-      const result = await movieService.delete(id._id);
+      const result = await moviesController.delete(id._id);
 
       //Assert
       expect(result).toBeUndefined();
@@ -565,7 +499,7 @@ describe('MovieService', () => {
       jest.spyOn(movieService, 'delete').mockRejectedValueOnce(new Error());
 
       //Assert
-      expect(movieService.delete(id._id)).rejects.toThrowError();
+      expect(moviesController.delete(id._id)).rejects.toThrowError();
     });
   });
 
@@ -578,7 +512,7 @@ describe('MovieService', () => {
   //     };
 
   //     // Act
-  //     const result = await movieService.findAndPaginate(body.limit, body.skip);
+  //     const result = await moviesController.findAndPaginate(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
@@ -591,7 +525,7 @@ describe('MovieService', () => {
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = {
+  //     const body: Pages = {
   //       limit: 15,
   //       skip: 1,
   //     };
@@ -602,9 +536,7 @@ describe('MovieService', () => {
   //       .mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(
-  //       movieService.findAndPaginate(body.limit, body.skip)
-  //     ).rejects.toThrowError();
+  //     expect(moviesController.findAndPaginate(body)).rejects.toThrowError();
   //   });
   // });
 
@@ -613,7 +545,7 @@ describe('MovieService', () => {
   //     // Arrange
 
   //     // Act
-  //     const result = await movieService.findAndCount();
+  //     const result = await moviesController.findAndCount();
 
   //     // Assert
   //     expect(typeof result).toEqual('number');
@@ -628,27 +560,27 @@ describe('MovieService', () => {
   //       .mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.findAndCount()).rejects.toThrowError();
+  //     expect(moviesController.findAndCount()).rejects.toThrowError();
   //   });
   // });
 
   // describe('getCategory', () => {
   //   it('Retorna lista de filmes por categoria', async () => {
   //     // Arrange
-  //     const body = 'Aventura';
+  //     const body = { category: 'Aventura' };
 
   //     // Act
-  //     const result = await movieService.getCategory(body);
+  //     const result = await moviesController.getCategory(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
   //     expect(movieService.getCategory).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getCategory).toHaveBeenCalledWith(body);
+  //     expect(movieService.getCategory).toHaveBeenCalledWith(body.category);
   //   });
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = 'Aventura';
+  //     const body = { category: 'Aventura' };
 
   //     // Arrange
   //     jest
@@ -656,27 +588,27 @@ describe('MovieService', () => {
   //       .mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.getCategory(body)).rejects.toThrowError();
+  //     expect(moviesController.getCategory(body)).rejects.toThrowError();
   //   });
   // });
 
   // describe('getDirectors', () => {
   //   it('Retorna lista de filmes de um diretor', async () => {
   //     // Arrange
-  //     const body = 'Tarantino';
+  //     const body = { director: 'Tarantino' };
 
   //     // Act
-  //     const result = await movieService.getDirectors(body);
+  //     const result = await moviesController.getDirectors(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
   //     expect(movieService.getDirectors).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getDirectors).toHaveBeenCalledWith(body);
+  //     expect(movieService.getDirectors).toHaveBeenCalledWith(body.director);
   //   });
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = 'Tarantino';
+  //     const body = { director: 'Tarantino' };
 
   //     // Arrange
   //     jest
@@ -684,85 +616,85 @@ describe('MovieService', () => {
   //       .mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.getDirectors(body)).rejects.toThrowError();
+  //     expect(moviesController.getDirectors(body)).rejects.toThrowError();
   //   });
   // });
 
   // describe('getCast', () => {
   //   it('Retorna lista de filmes de um ator', async () => {
   //     // Arrange
-  //     const body = 'Brad Pitt';
+  //     const body = { actor: 'Brad Pitt' };
 
   //     // Act
-  //     const result = await movieService.getCast(body);
+  //     const result = await moviesController.getCast(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
   //     expect(movieService.getCast).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getCast).toHaveBeenCalledWith(body);
+  //     expect(movieService.getCast).toHaveBeenCalledWith(body.actor);
   //   });
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = 'Brad Pitt';
+  //     const body = { actor: 'Brad Pitt' };
 
   //     // Arrange
   //     jest.spyOn(movieService, 'getCast').mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.getCast(body)).rejects.toThrowError();
+  //     expect(moviesController.getCast(body)).rejects.toThrowError();
   //   });
   // });
 
   // describe('getLetter', () => {
   //   it('Retorna lista de filmes pela letra de busca', async () => {
   //     // Arrange
-  //     const body = 'E';
+  //     const body = { letter: 'E' };
 
   //     // Act
-  //     const result = await movieService.getLetter(body);
+  //     const result = await moviesController.getLetter(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
   //     expect(movieService.getLetter).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getLetter).toHaveBeenCalledWith(body);
+  //     expect(movieService.getLetter).toHaveBeenCalledWith(body.letter);
   //   });
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = 'E';
+  //     const body = { letter: 'E' };
 
   //     // Arrange
   //     jest.spyOn(movieService, 'getLetter').mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.getLetter(body)).rejects.toThrowError();
+  //     expect(moviesController.getLetter(body)).rejects.toThrowError();
   //   });
   // });
 
   // describe('getByYear', () => {
   //   it('Retorna lista de filmes pelo ano de lançamento', async () => {
   //     // Arrange
-  //     const body = 2020;
+  //     const body = { year: 2020 };
 
   //     // Act
-  //     const result = await movieService.getByYear(body);
+  //     const result = await moviesController.getByYear(body);
 
   //     // Assert
   //     expect(result).toEqual(movie);
   //     expect(movieService.getByYear).toHaveBeenCalledTimes(1);
-  //     expect(movieService.getByYear).toHaveBeenCalledWith(body);
+  //     expect(movieService.getByYear).toHaveBeenCalledWith(body.year);
   //   });
 
   //   it('should throw an exception', () => {
   //     // Arrange
-  //     const body = 2020;
+  //     const body = { year: 2020 };
 
   //     // Arrange
   //     jest.spyOn(movieService, 'getByYear').mockRejectedValueOnce(new Error());
 
   //     // Assert
-  //     expect(movieService.getByYear(body)).rejects.toThrowError();
+  //     expect(moviesController.getByYear(body)).rejects.toThrowError();
   //   });
   // });
 });
